Extract creation of Photo* instances into separate SingletonManager
Removes any singleton characteristics from Photo{Manager,Factory},
and explicitly allow instantiation.
Uses dependency injection to provide PhotoManager with a PhotoFactory.

This allows for creating new objects in unit tests (which are in the
same package and have access to package-local constructors) while
forcing external users to get the single instances through the static
methods of the new SingletonManager.

Also, this makes it a lot easier to use other (derived) classes instead
of Photo{Manager,Factory} in the future. This could even be configured
during runtime, SingletonManager is factory-like in that relation.